moved routes into components folder
