+++ src/components/routes/Home.js
@@ -8,12 +8,12 @@ import {
   getDocs,
   endBefore,
 } from "firebase/firestore";
-import { db } from "../API/firebase";
-import { useAuth } from "../Auth/AuthContext";
-import ImageGridMasonry from "../components/ImgGrids/ImageGridMasonry";
-import Modal from "../components/Modal";
-import Title from "../components/Title";
-import HomeCards from "../components/UI/HomeCards";
+import { db } from "../../API/firebase";
+import { useAuth } from "../../Auth/AuthContext";
+import ImageGridMasonry from "../ImgGrids/ImageGridMasonry";
+import Modal from "../UI/Modal";
+import Title from "../Title";
+import HomeCards from "../UI/HomeCards";
 import Pagination from "@mui/material/Pagination";
 import Stack from "@mui/material/Stack";
 import Box from "@mui/material/Box";
