+++ src/components/routes/Login.js
@@ -1,8 +1,8 @@
 import React from "react";
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import { useAuth } from "../Auth/AuthContext";
-import LoadingCircle from "../components/UI/LoadingCircle";
+import { useAuth } from "../../Auth/AuthContext";
+import LoadingCircle from "../UI/LoadingCircle";
 import Avatar from "@mui/material/Avatar";
 import Button from "@mui/material/Button";
 import CssBaseline from "@mui/material/CssBaseline";
